+++ common/src/main/java/earth/terrarium/athena/mixins/ModelManagerMixin.java
@@ -1,14 +1,16 @@
 package earth.terrarium.athena.mixins;
 
-import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import earth.terrarium.athena.impl.loading.AthenaResourceLoader;
+import net.minecraft.client.renderer.block.model.BlockModel;
 import net.minecraft.client.resources.model.ModelManager;
-import net.minecraft.server.packs.resources.PreparableReloadListener;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
-import net.minecraft.util.profiling.ProfilerFiller;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
 
+import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 
@@ -16,17 +18,13 @@ import java.util.concurrent.Executor;
 public class ModelManagerMixin {
 
     // Force load our definitions before loading the vanilla ones
-    @WrapMethod(method = "reload")
-    private CompletableFuture<Void> wrapReload(
-            PreparableReloadListener.PreparationBarrier preparationBarrier,
+    @WrapOperation(method = "reload", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/resources/model/ModelManager;loadBlockModels(Lnet/minecraft/server/packs/resources/ResourceManager;Ljava/util/concurrent/Executor;)Ljava/util/concurrent/CompletableFuture;"))
+    private CompletableFuture<Map<ResourceLocation, BlockModel>> wrapReload(
             ResourceManager resourceManager,
-            ProfilerFiller profilerFiller,
-            ProfilerFiller profilerFiller2,
             Executor executor,
-            Executor executor2,
-            Operation<CompletableFuture<Void>> original
+            Operation<CompletableFuture<Map<ResourceLocation, BlockModel>>> original
     ) {
-        return AthenaResourceLoader.INSTANCE.reload(preparationBarrier, resourceManager, profilerFiller, profilerFiller2, executor, executor2)
-                .thenCompose(v -> original.call(preparationBarrier, resourceManager, profilerFiller, profilerFiller2, executor, executor2));
+        return CompletableFuture.runAsync(() -> AthenaResourceLoader.reload(resourceManager), executor)
+                .thenCompose((aVoid) -> original.call(resourceManager, executor));
     }
 }